+++ changes.md
@@ -3,6 +3,9 @@
 ## Changes in master
 1. Support of X axis with custom labels like:
       http://web-docs.gsi.de/~linev/js/dev/index.htm?nobrowser&json=../files/hist_xlabels.json
+2. Extend functionality of JSROOT.addDrawFunc() function. One could register type-specific
+   `make_request` and `after_request` functions; `icon`, `prereq`, `script`, `monitor` properties.
+   This let add more custom elements to the generic gui, implemented with JSROOT.HierarchyPainter   
 
 
 ## Changes in 3.6